fix/update-person-behavior

+++ src/modules/people/people.controller.ts
@@ -9,12 +9,13 @@ import {
   Request,
   UseGuards,
   Query,
+  UnauthorizedException,
+  ForbiddenException,
 } from '@nestjs/common';
 import { PeopleService } from './people.service';
 import { CreatePersonDto } from './dto/create-person.dto';
 import { UpdatePersonDto } from './dto/update-person.dto';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
-import { AdminGuard } from '../auth/admin-auth.guard';
 
 @Controller('people')
 export class PeopleController {
@@ -25,29 +26,57 @@ export class PeopleController {
     return this.peopleService.create(createPersonDto);
   }
 
-  @UseGuards(AdminGuard, JwtAuthGuard)
+  @UseGuards(JwtAuthGuard)
   @Get()
   findAll(@Query() query: any, @Request() req) {
     console.log('-------', req.user, '------');
     const { page, limit } = req.query;
-    return this.peopleService.findAll(page, limit, query);
+    if (req.user.admin == true) {
+      return this.peopleService.findAll(page, limit, query);
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 
   @UseGuards(JwtAuthGuard)
   @Get(':id')
-  findOne(@Param('id') id: string) {
-    return this.peopleService.findOne(id);
+  findOne(@Param('id') id: string, @Request() req) {
+    if (req.user.admin == true || id == req.user.id) {
+      return this.peopleService.findOne(id);
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 
   @UseGuards(JwtAuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updatePersonDto: UpdatePersonDto) {
-    return this.peopleService.update(id, updatePersonDto);
+  update(
+    @Param('id') id: string,
+    @Body() updatePersonDto: UpdatePersonDto,
+    @Request() req,
+  ) {
+    if (req.user.admin == true || id == req.user.id) {
+      return this.peopleService.update(id, updatePersonDto);
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 
   @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string) {
-    return this.peopleService.remove(id);
+  remove(@Param('id') id: string, @Request() req) {
+    if (req.user.admin == true || id == req.user.id) {
+      return this.peopleService.remove(id);
+    } else {
+      throw new ForbiddenException(
+        'Você não tem permissão para acessar esta rota',
+      );
+    }
   }
 }
